Cloud/requirements.txt
functions-framework
google-cloud-bigquery
google-cloud-pubsub
paho-mqtt 
requests
